Now using Python 3

+++ pygrader.py
@@ -1,60 +1,49 @@
 import argparse
-import os, sys, subprocess, shutil, time
+import os, sys, subprocess, shutil, time, glob, tempfile
 
+if sys.version_info[:3] < (3, 4, 0):
+	print('Requires Python >= 3.4.0')
+	sys.exit(1)
 
 parser = argparse.ArgumentParser()
 parser.add_argument('executable')
-parser.add_argument('--tests_dir', default='tests/')
+parser.add_argument('--tests', default='tests')
 parser.add_argument('--basename')
 parser.add_argument('--checker')
+parser.add_argument('--gen-solutions', action='store_true')
 parser.add_argument('--stdio', action='store_true')
 args = parser.parse_args()
 
 executable = args.executable
-
 if not os.path.isfile(executable):
-	print "Executable does not exist"
+	print("Executable '%s' does not exist" % os.path.abspath(executable))
 	raise SystemExit
 
-testsdir = os.path.join(os.path.dirname(os.path.abspath(executable)), args.tests_dir)
+testsdir = os.path.abspath(args.tests)
 if not os.path.isdir(testsdir):
-	print "Tests dir does not exist"
+	print("Tests directory '%s' does not exist" % testsdir)
 	raise SystemExit
-
-print 'Tests dir =', testsdir
 
 progname = args.basename or os.path.splitext(os.path.basename(executable))[0]
 
 test_file_in = progname + '.in'
 test_file_out = progname + '.out'
 
-print 'In/out files =', test_file_in, '/', test_file_out
+print('In/out files =', test_file_in, '/', test_file_out)
 
-# Enumerate tests
-
-tests = []
-
-for fn in os.listdir(testsdir):
-	if os.path.isfile(os.path.join(testsdir, fn)):
-		filename, ext = os.path.splitext(os.path.basename(fn))
-		if ext == '.in':
-			solfile = os.path.join(testsdir, filename + '.sol')
-			if os.path.isfile(solfile):
-				tests.append(filename)
-
-print "Found", len(tests), "test cases"
-
-exec_path = os.path.dirname(os.path.abspath(executable))
 
 def check_output(input_fn, output_fn, solution_fn):
 	if args.checker:
-		checker_output = subprocess.check_output([
+		with tempfile.TemporaryFile() as f_err:
+			checker_output = subprocess.check_output([
 				args.checker,
-				'10',  # points
 				input_fn,
-				output_fn,
-				solution_fn]).decode('utf-8').splitlines()
-		return (checker_output[0] == '10', checker_output[1])
+				solution_fn,
+				output_fn],
+				stderr=f_err).decode('utf-8').splitlines()
+			f_err.seek(0)
+			err_msg = f_err.read().decode('utf-8').strip()
+			return (checker_output[0] == '1', err_msg)
 	else:
 		output_contents = open(output_fn, "r").read()
 		solution_contents = open(solution_fn, "r").read()
@@ -63,13 +52,29 @@ def check_output(input_fn, output_fn, solution_fn):
 		else:
 			return (False, 'Wrong Answer')
 
+
+# Enumerate tests
+
+tests = []
+for fn in glob.glob(glob.escape(testsdir) + '/*.in'):
+	if os.path.isfile(fn):
+		filename, ext = os.path.splitext(os.path.basename(fn))
+		if (not args.gen_solutions and
+			not os.path.isfile(os.path.join(testsdir, filename + '.sol'))):
+			continue
+		tests.append(filename)
+
+print("Found", len(tests), "test cases")
+
+exec_path = os.path.dirname(os.path.abspath(executable))
+
 for test_name in tests:
 	input_fn = os.path.join(testsdir, test_name + '.in')
 	input_copy_fn = os.path.join(exec_path, test_file_in)
 	output_fn = os.path.join(exec_path, test_file_out)
 	solution_fn = os.path.join(testsdir, test_name + '.sol')
 	
-	print test_name,
+	print(test_name, end=' ')
 	# Copy input file
 	shutil.copy(input_fn, input_copy_fn)
 	exec_args = {}
@@ -81,24 +86,33 @@ for test_name in tests:
 	# Execute program
 	p = subprocess.Popen(executable, **exec_args)
 	t1 = time.time()
-	p.wait()
+	killed = False
+	try:
+		p.wait(2)
+	except subprocess.TimeoutExpired:
+		p.kill()
+		p.wait()
+		killed = True
 	t2 = time.time()
-	print "%.3f secs" % (t2 - t1),
-	if (t2-t1) > 2.0:
-		print "(timeout)",
+	print("%.3f secs" % (t2 - t1), end=' ')
+	if killed:
+		print("(killed)", end=' ')
 	
 	if args.stdio:
 		exec_args['stdin'].close()
 		exec_args['stdout'].close()
 	
-	success, message = check_output(input_fn, output_fn, solution_fn)
-	print message
-	if success:
-		os.remove(output_fn)
-	else:
-		new_name = os.path.join(exec_path, test_name + '.wrong.out')
-		try: os.unlink(new_name)
-		except: pass
-		os.rename(output_fn, new_name)
 	os.remove(input_copy_fn)
+	if not os.path.isfile(output_fn):
+		print('Output file not found')
+	else:
+		success, message = check_output(input_fn, output_fn, solution_fn)
+		print(message)
+		if success:
+			os.remove(output_fn)
+		else:
+			new_name = os.path.join(exec_path, test_name + '.wrong.out')
+			try: os.unlink(new_name)
+			except: pass
+			os.rename(output_fn, new_name)
